# Use gzip from a pool and stream while encoding metrics