@glefloch Here the promised joint PR.