@glefloch 2018-10-20
This looks pretty neat! We're you able to measure memory and CPU gain?