@beorn7 2018-10-20
I'm about to try it out with our kube-state-metrics setup at SoundCloud. It spits out ~150MiB of metrics (uncompressed text format) on each scrape… It's for a cluster with about 10k pods. Will report back once I have results.