@glefloch 2018-10-21
Well, do you think this is only due to compression or due to the fact there is no more intermediate buffer?