@beorn7 2018-10-21
I compared compression to compression and no-compression to no-compression. Each time without this PR vs. with this PR. But I think the weird results were because my pod landed on a slow node. I redeployed the whole shebang, and now the metrics look more consistent. But we should take it as a reminder that on a cluster, things can vary a lot, not only due to different hardware of node but also different load state of nodes etc. So taking all the results with a grain of salt, I'd say that this PR reduces scrape duration and CPU usage by broadly 20% without compression and broadly 10% with compression. Since this is high-volume low-frequency scraping, the sync.Pool will have no real gain but leaving out the extra buffer is a net saving.  It makes sense that the net gain is lower with enabled compression as previously, we wrote the _compressed_ metrics into the extra buffer, which is much smaller than the uncompressed metrics.

To prove the use of the sync.Pool, I'll try a load test ASAP.