@glefloch 2018-10-22
@beorn7, this looks good to me. I have nothing to add.