@beorn7 2018-10-22
Load test of the "random" example with 5000 requests in 20s, median and 99th percentile latency, CPU usage,
* before, no gzip: 12ms, 72ms, 140%
* before, with gzip: 11ms, 76ms, 130%
* after, no gzip: 10ms, 77ms, 130%,
* before, with gzip: 9ms, 79ms, 120%

While a small improvement is visible, this is very close to noise.

Overall, I'd still say the change is worth it. @glefloch let me know if you have anything to add or change here. Otherwise, I'll also propagate this into the legacy HTTP code in the `prometheus` package before merging.