@beorn7 2018-10-22
GH seemed to have issues last night. I get a mail from travis that the test were passing after my final commit.