@beorn7 2018-10-22
FTR: comment repeats were happening because of a GH outage.

All tests green now. I'll merge.

Thanks everybody for contributing to the discussion and the code here and in #481 .