@beorn7 2018-10-22
Added a commit to have the same improvements in the deprecated HTTP stuff in the `prometheus` package.